third_party/blink/web_tests/external/wpt/webaudio/the-audio-api/the-delaynode-interface/delaynode-channel-count-1-expected.txt

This is a testharness.js-based test.
[FAIL] Test that DelayNode output channelCount matches that of the delayed input
  assert_equals: Initial mono at 0 expected 0 but got 0.25
Harness: the test ran to completion.
